+++ khd/enums.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class KernelBackend(Enum):
+    cuda = "cuda"
+    triton = "triton"
