+++ khd/kernels/__init__.py
@@ -3,6 +3,7 @@ from .add import (
     add_scalar_torch,
     add_scalar_triton,
     add_tensor_cuda,
+    add_tensor_generic,
     add_tensor_torch,
     add_tensor_triton,
 )
