+++ khd/kernels/add/__init__.py
@@ -1,2 +1,2 @@
 from .add_scalar import add_scalar_cuda, add_scalar_torch, add_scalar_triton
-from .add_tensor import add_tensor_cuda, add_tensor_torch, add_tensor_triton
+from .add_tensor import add_tensor_cuda, add_tensor_generic, add_tensor_torch, add_tensor_triton
